Fix race condition that happens when servers join
Previously if the new server joins before the old servers have
restarted, the whole job fails because the new server throws an
exception complaining that the queues are not the right size.

This commit fixes it by adding a check before a server joins.
This check makes sure everyone has the new cluster_spec before
building the graph, so no one builds a queue with the wrong
capacity. This requires adding an autoscaling client on each
of the servers and complex error handling to make sure no
additional race conditions are introduced.