Fix new worker bootstrapping problem
Previously, if a new worker doesn't have the same cluster spec
as everyone else then it can't join and the job will hang. This
commit adds the concept of a "master autoscaling server"
designated to be the first worker of every cluster. This master
holds the ground truth cluster spec that everyone uses.

One special case is the following. When a new worker joins,
the master may not know about the worker yet. In that case
the new worker shouldn't just blindly follow the master's
cluster spec because it doesn't include the new worker.
Instead, the new worker will block until the master's cluster
spec contains the new worker before proceeding.