Enable autoscaling for distributed mode
This commit adds a mechanism through which the user can launch
a single worker that can attach itself to an existing cluster.
This is done by setting the AUTOSCALING_MASTER_HOST_PORT
environment variable to the host port of the master autoscaling
server (first worker) when launching the worker. Obviously,
NUM_WORKERS should be 1 and NUM_PARAMETER_SERVERS should be 0.