Implement join_cluster: allow worker to self join
Previously every time we want to change the membership of a
cluster we have to manually send an add_worker message to each
machine to trigger a restart. Now the restart will be triggered
instead by the new worker when it comes alive and wants to join
the cluster.

This allows us to launch workers asynchronously and keep training
while the new worker is still coming up.