Synchronize with epochs + pending cluster specs
This commit gets rid of the initialization lock, which was held
for a long time and became a common cause of deadlocks. Instead,
we introduce epochs, which is incremented every time we restart.
Building a graph requires everyone to be on the same epoch or
above.

This commit also introduces pending cluster specs, which are
applied only when restarting to ensure we don't advertise a new
cluster spec while running on an old graph.