Introduce autoscaling statuses to fix race conditions
This commit gets rid of the epochs previously introduced and
replaces it with autoscaling statuses. The goal is to make sure
everyone restarts in lockstep; no one can be more than one status
away from each other in the finite state machine. This solves
most of the race conditions that occur previously when starting
multiple workers at once.