Fix remove
We kept waiting for removed workers to be SYNCING, but they
actually already terminated. Thus we should add a TERMINATED
state. Note that this will fail if the process actually exits,
in which case even fetching the remote status will fail. We may
need to address this in the future when we handle fault tolerance.